make day7/main.rb into day7.rb and rename day7/ into intcode_master/ so that IntcodeProgram can be reused on the daily
